Beginning of Seqsee.pl. Reads a config file, gives precedence to command line switches, supports defauls, starts logging correctly if requested etc
git-svn-id: http://seqsee1.googlecode.com/svn/Seqsee/branches/summer05@165 19b06628-3642-0410-8fdd-33ae7300e995

+++ lib/SLog.pm
@@ -15,10 +15,13 @@ use List::Util;
 
 sub init{
     my ( $pack, $opts_ref ) = @_;
-    my $logging = $opts_ref->{logging} || 0;
-    my $sequence = $opts_ref->{sequence} || [];
-    my $seed = $opts_ref->{seed} || 0;
-    my $steps = $opts_ref->{steps} || 0;
+    unless (exists $opts_ref->{log}) {
+        die "Expected information about whether to log";
+    }
+    my $logging = $opts_ref->{log} || 0;
+    my $sequence = $opts_ref->{seq} || die "Expected to see a sequence";
+    my $seed = $opts_ref->{seed} || die "expected to see a seed";
+    my $steps = $opts_ref->{max_steps} || die "expected max  steps";
 
     $sequence = join(", ", @$sequence);
 
@@ -49,8 +52,11 @@ NOLOG
                            );
         $logger->info($msg);
     }
-    unlink "log/latest";
-    system "ln -s $SLog::filename log/latest";
+
+    if ($logging) {
+        unlink "log/latest";
+        system "ln -s $SLog::filename log/latest";
+    }
 }
 
 1;
